+++ packages/fat-enums/tests/Fixtures/ModelBackedTestEnum.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace ArtisanBuild\FatEnums\Tests\Fixtures;
 
-use ArtisanBuild\FatEnums\Tests\Fixtures\Models\DummyModel;
+use ArtisanBuild\FatEnums\Tests\Fixtures\DummyModel;
 use ArtisanBuild\FatEnums\Traits\DatabaseRecordsEnum;
 
 enum ModelBackedTestEnum: int
